# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /planning:
    get:
      operationId: listPlans
      tags: [planning]
      summary: List monthly plans
      parameters:
        - in: query
          name: managerId
          schema: { type: ["integer", "null"] }
        - in: query
          name: month
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [planning]
      summary: Create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /planning/{id}:
    patch:
      operationId: updatePlan
      tags: [planning]
      summary: Update plan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found
...
  /workspace/my-stats:
    get:
      operationId: getMyStats
      tags: [workspace]
      summary: Manager workspace stats for current user
      parameters:
        - in: query
          name: month
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceStats"
components:
  schemas:
...
    Plan:
      type: object
      required: [id, managerId, month, minPlan, targetPlan, maxPlan]
      properties:
        id: { type: integer }
        managerId: { type: integer }
        managerName: { type: ["string", "null"] }
        month: { type: string }
        product: { type: ["string", "null"] }
        minPlan: { type: number }
        targetPlan: { type: number }
        maxPlan: { type: number }
        createdAt: { type: string }
    PlanInput:
      type: object
      required: [managerId, month, minPlan, targetPlan, maxPlan]
      properties:
        managerId: { type: integer }
        month: { type: string }
        product: { type: string }
        minPlan: { type: number }
        targetPlan: { type: number }
        maxPlan: { type: number }
    PlanUpdate:
      type: object
      properties:
        minPlan: { type: number }
        targetPlan: { type: number }
        maxPlan: { type: number }
        product: { type: string }
...
    WorkspaceStats:
      type: object
      required: [managerId, managerName, openDeals, todayTasks, currentBonus, amountToTarget, currentMultiplier, netProfit, targetPlan, minPlan]
      properties:
        managerId: { type: integer }
        managerName: { type: string }
        openDeals: { type: integer }
        todayTasks: { type: integer }
        currentBonus: { type: number }
        amountToTarget: { type: number }
        currentMultiplier: { type: number }
        netProfit: { type: number }
        targetPlan: { type: number }
        minPlan: { type: number }
        planProgress: { type: number }