# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  managerId: number;
  /** @nullable */
  managerName?: string | null;
  month: string;
  /** @nullable */
  product?: string | null;
  minPlan: number;
  targetPlan: number;
  maxPlan: number;
  createdAt?: string;
}