# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanUpdate {
  minPlan?: number;
  targetPlan?: number;
  maxPlan?: number;
  product?: string;
}