# lib/api-zod/src/generated/types/workspaceStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspaceStats {
  managerId: number;
  managerName: string;
  openDeals: number;
  todayTasks: number;
  currentBonus: number;
  amountToTarget: number;
  currentMultiplier: number;
  netProfit: number;
  targetPlan: number;
  minPlan: number;
  planProgress?: number;
}